webserver/requirements.txt
django==4.2.11
djangorestframework==3.15.1
django-filter==24.2
numpy==1.24.4
